test: 질문 식별자로 삭제되지 않은 답변 조회 테스트 작성
